Refactor homepage sections into components
